Prep for reporting location of template compilation errors.
Step 3: unroll main substitution into an explicit loop in
which we can track where each subst. occurs. With this and
the line-break index we build up front, we'll be able to
determine the source line and start/end line inded for
each replaced tag, and can then report that information
when a compilation failure (i.e. bad template syntax) is
detected.